Reorganise some class locations
