+++ SimpCity/ClassDiagram.cd
@@ -30,7 +30,8 @@
     <Position X="0.75" Y="2" Width="1.5" />
     <TypeIdentifier>
       <HashCode>AAIAAgAAAAAAAAAgAAAAIAAAAAAAAAAAAAAAAEAAAAA=</HashCode>
-      <FileName>CityGrid.cs</FileName>
+      <FileName>CityGridBuilding.cs</FileName>
+      <NewMemberFileName>CityGrid.cs</NewMemberFileName>
     </TypeIdentifier>
     <ShowAsAssociation>
       <Property Name="Grid" />
@@ -40,12 +41,12 @@
   <Class Name="SimpCity.CityGrid">
     <Position X="0.75" Y="4.75" Width="1.5" />
     <TypeIdentifier>
-      <HashCode>AAIEAAAABAIAAABAAAAAAEAAAEAgAAAAAAEAAAAAAAA=</HashCode>
+      <HashCode>AAIAAAAABAIAAABAAAAAAEAAAEAgAAAAAAEAAAAAAAA=</HashCode>
       <FileName>CityGrid.cs</FileName>
     </TypeIdentifier>
   </Class>
   <Class Name="SimpCity.buildings.Beach" Collapsed="true">
-    <Position X="0.75" Y="1" Width="1.5" />
+    <Position X="0.75" Y="0.75" Width="1.5" />
     <TypeIdentifier>
       <HashCode>AAAAAAAAAAAAAAAAAAAgAAQAAAAAAAAAAAAAAAAAAAA=</HashCode>
       <FileName>buildings\Beach.cs</FileName>
