+++ SimpCity/Enums.cs
@@ -1,0 +1,8 @@
+﻿namespace SimpCity {
+    /// <summary>
+    /// Enumerates the different building types.
+    /// </summary>
+    public enum BuildingTypes {
+        Beach,
+    }
+}
